+++ contracts/hackatom/src/contract.rs
@@ -150,11 +150,11 @@ fn query_verifier<S: Storage, A: Api>(deps: &Extern<S, A>) -> Result<Vec<u8>> {
 mod tests {
     use super::*;
     // import trait ReadonlyStorage to get access to read
-    use cosmwasm::{coin, dependencies, mock_env, transactional_deps, ReadonlyStorage};
+    use cosmwasm::{coin, mock_dependencies, mock_env, transactional_deps, ReadonlyStorage};
 
     #[test]
     fn proper_initialization() {
-        let mut deps = dependencies(20);
+        let mut deps = mock_dependencies(20);
 
         let verifier = HumanAddr(String::from("verifies"));
         let beneficiary = HumanAddr(String::from("benefits"));
@@ -181,7 +181,7 @@ mod tests {
 
     #[test]
     fn init_and_query() {
-        let mut deps = dependencies(20);
+        let mut deps = mock_dependencies(20);
 
         let verifier = HumanAddr(String::from("verifies"));
         let beneficiary = HumanAddr(String::from("benefits"));
@@ -202,7 +202,7 @@ mod tests {
 
     #[test]
     fn checkpointing_works_on_contract() {
-        let mut deps = dependencies(20);
+        let mut deps = mock_dependencies(20);
 
         let verifier = HumanAddr(String::from("verifies"));
         let beneficiary = HumanAddr(String::from("benefits"));
@@ -234,7 +234,7 @@ mod tests {
 
     #[test]
     fn proper_handle() {
-        let mut deps = dependencies(20);
+        let mut deps = mock_dependencies(20);
 
         // initialize the store
         let verifier = HumanAddr(String::from("verifies"));
@@ -279,7 +279,7 @@ mod tests {
 
     #[test]
     fn failed_handle() {
-        let mut deps = dependencies(20);
+        let mut deps = mock_dependencies(20);
 
         // initialize the store
         let verifier = HumanAddr(String::from("verifies"));
@@ -320,7 +320,7 @@ mod tests {
     #[test]
     #[should_panic(expected = "This page intentionally faulted")]
     fn handle_panic() {
-        let mut deps = dependencies(20);
+        let mut deps = mock_dependencies(20);
 
         // initialize the store
         let verifier = HumanAddr(String::from("verifies"));