+++ packages/vm/src/cache.rs
@@ -133,7 +133,7 @@ where
 mod test {
     use super::*;
     use crate::calls::{call_handle, call_init};
-    use cosmwasm::{coin, dependencies, mock_env, MockApi, MockStorage};
+    use cosmwasm::{coin, mock_dependencies, mock_env, MockApi, MockStorage};
     use tempfile::TempDir;
 
     static TESTING_GAS_LIMIT: u64 = 400_000;
@@ -189,7 +189,7 @@ mod test {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT_0_7).unwrap();
-        let deps = dependencies(20);
+        let deps = mock_dependencies(20);
         let _instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
         assert_eq!(cache.stats.hits_instance, 0);
         assert_eq!(cache.stats.hits_module, 1);
@@ -201,9 +201,9 @@ mod test {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT_0_7).unwrap();
-        let deps1 = dependencies(20);
-        let deps2 = dependencies(20);
-        let deps3 = dependencies(20);
+        let deps1 = mock_dependencies(20);
+        let deps2 = mock_dependencies(20);
+        let deps3 = mock_dependencies(20);
         let instance1 = cache.get_instance(&id, deps1, TESTING_GAS_LIMIT).unwrap();
         cache.store_instance(&id, instance1);
         let instance2 = cache.get_instance(&id, deps2, TESTING_GAS_LIMIT).unwrap();
@@ -220,7 +220,7 @@ mod test {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT_0_7).unwrap();
-        let deps = dependencies(20);
+        let deps = mock_dependencies(20);
         let mut instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
 
         // run contract
@@ -238,7 +238,7 @@ mod test {
         let tmp_dir = TempDir::new().unwrap();
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT_0_7).unwrap();
-        let deps = dependencies(20);
+        let deps = mock_dependencies(20);
         let mut instance = cache.get_instance(&id, deps, TESTING_GAS_LIMIT).unwrap();
 
         // init contract
@@ -268,8 +268,8 @@ mod test {
         let id = cache.save_wasm(CONTRACT_0_7).unwrap();
 
         // these differentiate the two instances of the same contract
-        let deps1 = dependencies(20);
-        let deps2 = dependencies(20);
+        let deps1 = mock_dependencies(20);
+        let deps2 = mock_dependencies(20);
 
         // init instance 1
         let mut instance = cache.get_instance(&id, deps1, TESTING_GAS_LIMIT).unwrap();
@@ -325,8 +325,8 @@ mod test {
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT_0_7).unwrap();
 
-        let deps1 = dependencies(20);
-        let deps2 = dependencies(20);
+        let deps1 = mock_dependencies(20);
+        let deps2 = mock_dependencies(20);
 
         // Init from module cache
         let mut instance1 = cache.get_instance(&id, deps1, TESTING_GAS_LIMIT).unwrap();
@@ -357,8 +357,8 @@ mod test {
         let mut cache = unsafe { CosmCache::new(tmp_dir.path(), 10).unwrap() };
         let id = cache.save_wasm(CONTRACT_0_7).unwrap();
 
-        let deps1 = dependencies(20);
-        let deps2 = dependencies(20);
+        let deps1 = mock_dependencies(20);
+        let deps2 = mock_dependencies(20);
 
         // Init from module cache
         let mut instance1 = cache.get_instance(&id, deps1, 10).unwrap();