+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Export all symbols at top level (e.g.
   `use cosmwasm::traits::{Api, Storage};` + `use cosmwasm::encoding::Binary;`
   becomes `use cosmwasm::{Api, Binary, Storage};`).
+- Rename `cosmwasm::dependencies` -> `cosmwasm::mock_dependencies` to
+  differentiate between testing and production `External`.
 
 ## 0.7.2 (2020-03-23)